+++ virotour/flask/app/tests/conftest.py
@@ -1,0 +1,14 @@
+import os
+
+import pytest
+from app import create_app, db
+
+@pytest.fixture
+def app():
+    app = create_app()
+    with app.app_context():
+        db.create_all()
+        yield app   # Note that we changed return for yield, see below for why
+        db.session.remove()
+        db.drop_all()
+    return app
