+++ virotour/flask/requirements.txt
@@ -1,7 +1,6 @@
 click==8.1.3
 colorama==0.4.6
-Flask==2.2.3
-Flask-SQLAlchemy==3.0.3
+Flask-Testing==0.8.1
 Flask-WTF==1.1.1
 greenlet==2.0.2
 importlib-metadata==6.0.0
